feat: Extract IndexContext into State component
